+++ src/acker.rs
@@ -1,8 +1,15 @@
 use crate::{
+    error_holder::ErrorHolder,
     internal_rpc::InternalRPCHandle,
     options::{BasicAckOptions, BasicNackOptions, BasicRejectOptions},
+    protocol::{AMQPError, AMQPSoftError},
     types::{ChannelId, DeliveryTag},
-    Promise, PromiseResolver, Result,
+    Error, Promise, PromiseResolver, Result,
+};
+
+use std::sync::{
+    atomic::{AtomicBool, Ordering},
+    Arc,
 };
 
 #[derive(Default, Debug, Clone)]
@@ -10,6 +17,8 @@ pub struct Acker {
     channel_id: ChannelId,
     delivery_tag: DeliveryTag,
     internal_rpc: Option<InternalRPCHandle>,
+    error: Option<ErrorHolder>,
+    used: Arc<AtomicBool>,
 }
 
 impl Acker {
@@ -17,41 +26,76 @@ impl Acker {
         channel_id: ChannelId,
         delivery_tag: DeliveryTag,
         internal_rpc: Option<InternalRPCHandle>,
+        error: Option<ErrorHolder>,
     ) -> Self {
         Self {
             channel_id,
             delivery_tag,
             internal_rpc,
+            error,
+            used: Arc::default(),
         }
     }
 
+    // FIXME: consume self and drop used
     pub async fn ack(&self, options: BasicAckOptions) -> Result<()> {
         self.rpc(|internal_rpc, resolver| {
-            internal_rpc.basic_ack(self.channel_id, self.delivery_tag, options, resolver)
+            internal_rpc.basic_ack(
+                self.channel_id,
+                self.delivery_tag,
+                options,
+                resolver,
+                self.error.clone(),
+            )
         })
         .await
     }
 
+    // FIXME: consume self and drop used
     pub async fn nack(&self, options: BasicNackOptions) -> Result<()> {
         self.rpc(|internal_rpc, resolver| {
-            internal_rpc.basic_nack(self.channel_id, self.delivery_tag, options, resolver)
+            internal_rpc.basic_nack(
+                self.channel_id,
+                self.delivery_tag,
+                options,
+                resolver,
+                self.error.clone(),
+            )
         })
         .await
     }
 
+    // FIXME: consume self and drop used
     pub async fn reject(&self, options: BasicRejectOptions) -> Result<()> {
         self.rpc(|internal_rpc, resolver| {
-            internal_rpc.basic_reject(self.channel_id, self.delivery_tag, options, resolver)
+            internal_rpc.basic_reject(
+                self.channel_id,
+                self.delivery_tag,
+                options,
+                resolver,
+                self.error.clone(),
+            )
         })
         .await
     }
 
     async fn rpc<F: Fn(&InternalRPCHandle, PromiseResolver<()>)>(&self, f: F) -> Result<()> {
-        let (promise, resolver) = Promise::new();
-        if let Some(internal_rpc) = self.internal_rpc.as_ref() {
-            f(internal_rpc, resolver);
+        if self.used.swap(true, Ordering::SeqCst) {
+            return Err(Error::ProtocolError(AMQPError::new(
+                AMQPSoftError::PRECONDITIONFAILED.into(),
+                "Attempted to use an already used Acker".into(),
+            )));
         }
-        promise.await
+        if let Some(error) = self.error.as_ref() {
+            error.check()?;
+        }
+        if let Some(internal_rpc) = self.internal_rpc.as_ref() {
+            let (promise, resolver) = Promise::new();
+            f(internal_rpc, resolver);
+            promise.await
+        } else {
+            Ok(())
+        }
     }
 }