+++ src/acknowledgement.rs
@@ -1,8 +1,9 @@
 use crate::{
+    id_sequence::IdSequence,
     protocol::{AMQPError, AMQPSoftError},
     publisher_confirm::{Confirmation, PublisherConfirm},
     returned_messages::ReturnedMessages,
-    types::{ChannelId, DeliveryTag},
+    types::DeliveryTag,
     Error, Promise, Result,
 };
 use parking_lot::Mutex;
@@ -11,6 +12,7 @@ use std::{
     fmt,
     sync::Arc,
 };
+use tracing::trace;
 
 #[derive(Clone)]
 pub(crate) struct Acknowledgements(Arc<Mutex<Inner>>);
@@ -19,16 +21,15 @@ type AMQPResult = std::result::Result<(), AMQPError>;
 type ConfirmationBroadcaster = pinky_swear::PinkyBroadcaster<Result<Confirmation>>;
 
 impl Acknowledgements {
-    pub(crate) fn new(returned_messages: ReturnedMessages) -> Self {
-        Self(Arc::new(Mutex::new(Inner::new(returned_messages))))
+    pub(crate) fn new(channel_id: u16, returned_messages: ReturnedMessages) -> Self {
+        Self(Arc::new(Mutex::new(Inner::new(
+            channel_id,
+            returned_messages,
+        ))))
     }
 
-    pub(crate) fn register_pending(
-        &self,
-        delivery_tag: DeliveryTag,
-        channel_id: ChannelId,
-    ) -> PublisherConfirm {
-        self.0.lock().register_pending(delivery_tag, channel_id)
+    pub(crate) fn register_pending(&self) -> PublisherConfirm {
+        self.0.lock().register_pending()
     }
 
     pub(crate) async fn get_last_pending(&self) -> Option<crate::Result<Confirmation>> {
@@ -36,12 +37,12 @@ impl Acknowledgements {
         Some(promise.await)
     }
 
-    pub(crate) fn ack(&self, delivery_tag: DeliveryTag, channel_id: ChannelId) -> AMQPResult {
-        self.0.lock().drop_pending(delivery_tag, true, channel_id)
+    pub(crate) fn ack(&self, delivery_tag: DeliveryTag) -> AMQPResult {
+        self.0.lock().drop_pending(delivery_tag, true)
     }
 
-    pub(crate) fn nack(&self, delivery_tag: DeliveryTag, channel_id: ChannelId) -> AMQPResult {
-        self.0.lock().drop_pending(delivery_tag, false, channel_id)
+    pub(crate) fn nack(&self, delivery_tag: DeliveryTag) -> AMQPResult {
+        self.0.lock().drop_pending(delivery_tag, false)
     }
 
     pub(crate) fn ack_all_pending(&self) {
@@ -52,28 +53,16 @@ impl Acknowledgements {
         self.0.lock().drop_all(false);
     }
 
-    pub(crate) fn ack_all_before(
-        &self,
-        delivery_tag: DeliveryTag,
-        channel_id: ChannelId,
-    ) -> AMQPResult {
-        self.0
-            .lock()
-            .complete_pending_before(delivery_tag, true, channel_id)
+    pub(crate) fn ack_all_before(&self, delivery_tag: DeliveryTag) -> AMQPResult {
+        self.0.lock().complete_pending_before(delivery_tag, true)
     }
 
-    pub(crate) fn nack_all_before(
-        &self,
-        delivery_tag: DeliveryTag,
-        channel_id: ChannelId,
-    ) -> AMQPResult {
-        self.0
-            .lock()
-            .complete_pending_before(delivery_tag, false, channel_id)
+    pub(crate) fn nack_all_before(&self, delivery_tag: DeliveryTag) -> AMQPResult {
+        self.0.lock().complete_pending_before(delivery_tag, false)
     }
 
-    pub(crate) fn on_channel_error(&self, channel_id: ChannelId, error: Error) {
-        self.0.lock().on_channel_error(channel_id, error);
+    pub(crate) fn on_channel_error(&self, error: Error) {
+        self.0.lock().on_channel_error(error);
     }
 }
 
@@ -82,6 +71,7 @@ impl fmt::Debug for Acknowledgements {
         let mut debug = f.debug_struct("Acknowledgements");
         if let Some(inner) = self.0.try_lock() {
             debug
+                .field("delivery_tag", &inner.delivery_tag)
                 .field("returned_messages", &inner.returned_messages)
                 .field("pending", &inner.pending.keys());
         }
@@ -90,35 +80,37 @@ impl fmt::Debug for Acknowledgements {
 }
 
 struct Inner {
+    channel_id: u16,
+    delivery_tag: IdSequence<DeliveryTag>,
     last: Option<(DeliveryTag, Promise<Confirmation>)>,
-    pending: HashMap<DeliveryTag, (ChannelId, ConfirmationBroadcaster)>,
+    pending: HashMap<DeliveryTag, ConfirmationBroadcaster>,
     returned_messages: ReturnedMessages,
 }
 
 impl Inner {
-    fn new(returned_messages: ReturnedMessages) -> Self {
+    fn new(channel_id: u16, returned_messages: ReturnedMessages) -> Self {
         Self {
+            channel_id,
+            delivery_tag: IdSequence::new(false),
             last: None,
             pending: HashMap::default(),
             returned_messages,
         }
     }
 
-    fn register_pending(
-        &mut self,
-        delivery_tag: DeliveryTag,
-        channel_id: ChannelId,
-    ) -> PublisherConfirm {
+    fn register_pending(&mut self) -> PublisherConfirm {
+        let delivery_tag = self.delivery_tag.next();
+        trace!("Publishing with delivery_tag {}", delivery_tag);
         let broadcaster = ConfirmationBroadcaster::default();
         let promise =
             PublisherConfirm::new(broadcaster.subscribe(), self.returned_messages.clone());
         if let Some((delivery_tag, promise)) = self.last.take() {
-            if let Some((_, broadcaster)) = self.pending.get(&delivery_tag) {
+            if let Some(broadcaster) = self.pending.get(&delivery_tag) {
                 broadcaster.unsubscribe(promise);
             }
         }
         self.last = Some((delivery_tag, broadcaster.subscribe()));
-        self.pending.insert(delivery_tag, (channel_id, broadcaster));
+        self.pending.insert(delivery_tag, broadcaster);
         promise
     }
 
@@ -132,45 +124,37 @@ impl Inner {
     }
 
     fn drop_all(&mut self, success: bool) {
-        for (_, resolver) in self
+        for resolver in self
             .pending
             .drain()
-            .map(|tup| tup.1)
-            .collect::<Vec<(ChannelId, ConfirmationBroadcaster)>>()
+            .map(|(_, resolver)| resolver)
+            .collect::<Vec<_>>()
         {
             self.complete_pending(success, resolver);
         }
     }
 
-    fn drop_pending(
-        &mut self,
-        delivery_tag: DeliveryTag,
-        success: bool,
-        channel_id: ChannelId,
-    ) -> AMQPResult {
-        if let Some((_, resolver)) = self.pending.remove(&delivery_tag) {
+    fn drop_pending(&mut self, delivery_tag: DeliveryTag, success: bool) -> AMQPResult {
+        if let Some(resolver) = self.pending.remove(&delivery_tag) {
             self.complete_pending(success, resolver);
             Ok(())
         } else {
             Err(AMQPError::new(
                 AMQPSoftError::PRECONDITIONFAILED.into(),
                 format!(
-                    "invalid {} received for non-existent delivery_tag {} on channel {}",
+                    "invalid {} received for inexistant delivery_tag {} on channel {}, current is {}, was expecting one of {:?}",
                     if success { "ack" } else { "nack" },
                     delivery_tag,
-                    channel_id
+                    self.channel_id,
+                    self.delivery_tag.current(),
+                    self.pending.keys()
                 )
                 .into(),
             ))
         }
     }
 
-    fn complete_pending_before(
-        &mut self,
-        delivery_tag: DeliveryTag,
-        success: bool,
-        channel_id: ChannelId,
-    ) -> AMQPResult {
+    fn complete_pending_before(&mut self, delivery_tag: DeliveryTag, success: bool) -> AMQPResult {
         let mut res = Ok(());
         for tag in self
             .pending
@@ -179,22 +163,16 @@ impl Inner {
             .cloned()
             .collect::<HashSet<DeliveryTag>>()
         {
-            if let Err(err) = self.drop_pending(tag, success, channel_id) {
+            if let Err(err) = self.drop_pending(tag, success) {
                 res = Err(err);
             }
         }
         res
     }
 
-    fn on_channel_error(&mut self, channel_id: ChannelId, error: Error) {
-        for (_, resolver) in self
-            .pending
-            .values()
-            .filter(|(channel, _)| *channel == channel_id)
-        {
+    fn on_channel_error(&mut self, error: Error) {
+        for (_, resolver) in self.pending.drain() {
             resolver.swear(Err(error.clone()));
         }
-        self.pending
-            .retain(|_, (channel, _)| *channel != channel_id);
     }
 }