+++ src/consumer.rs
@@ -2,6 +2,7 @@ use crate::{
     channel_closer::ChannelCloser,
     consumer_canceler::ConsumerCanceler,
     consumer_status::{ConsumerState, ConsumerStatus},
+    error_holder::ErrorHolder,
     internal_rpc::InternalRPCHandle,
     message::{Delivery, DeliveryResult},
     options::BasicConsumeOptions,
@@ -189,6 +190,10 @@ impl Consumer {
         }
     }
 
+    pub(crate) fn error(&self) -> ErrorHolder {
+        self.inner.lock().error.clone()
+    }
+
     /// Gets the consumer tag.
     ///
     /// If no consumer tag was specified when obtaining the consumer from the channel,
@@ -273,6 +278,7 @@ struct ConsumerInner {
     deliveries_in: Sender<DeliveryResult>,
     deliveries_out: Receiver<DeliveryResult>,
     wakers: Wakers,
+    error: ErrorHolder,
     tag: ShortString,
     delegate: Option<Arc<Box<dyn ConsumerDelegate>>>,
     executor: Arc<dyn FullExecutor + Send + Sync>,
@@ -304,6 +310,7 @@ impl ConsumerInner {
             deliveries_in: sender,
             deliveries_out: receiver,
             wakers: Wakers::default(),
+            error: ErrorHolder::default(),
             tag: consumer_tag,
             delegate: None,
             executor,
@@ -381,6 +388,7 @@ impl ConsumerInner {
 
     fn set_error(&mut self, error: Error) {
         trace!(consumer_tag=%self.tag, "set_error");
+        self.error.set(error.clone());
         if let Some(delegate) = self.delegate.as_ref() {
             let delegate = delegate.clone();
             self.executor.spawn(delegate.on_new_delivery(Err(error)));