+++ src/internal_rpc.rs
@@ -1,6 +1,7 @@
 use crate::{
     channels::Channels,
     consumer_status::ConsumerStatus,
+    error_holder::ErrorHolder,
     options::{BasicAckOptions, BasicCancelOptions, BasicNackOptions, BasicRejectOptions},
     socket_state::SocketStateHandle,
     types::{ChannelId, DeliveryTag, Identifier, ReplyCode},
@@ -30,12 +31,14 @@ impl InternalRPCHandle {
         delivery_tag: DeliveryTag,
         options: BasicAckOptions,
         resolver: PromiseResolver<()>,
+        error: Option<ErrorHolder>,
     ) {
         self.send(InternalCommand::BasicAck(
             channel_id,
             delivery_tag,
             options,
             resolver,
+            error,
         ));
     }
 
@@ -45,12 +48,14 @@ impl InternalRPCHandle {
         delivery_tag: DeliveryTag,
         options: BasicNackOptions,
         resolver: PromiseResolver<()>,
+        error: Option<ErrorHolder>,
     ) {
         self.send(InternalCommand::BasicNack(
             channel_id,
             delivery_tag,
             options,
             resolver,
+            error,
         ));
     }
 
@@ -60,12 +65,14 @@ impl InternalRPCHandle {
         delivery_tag: DeliveryTag,
         options: BasicRejectOptions,
         resolver: PromiseResolver<()>,
+        error: Option<ErrorHolder>,
     ) {
         self.send(InternalCommand::BasicReject(
             channel_id,
             delivery_tag,
             options,
             resolver,
+            error,
         ));
     }
 
@@ -170,18 +177,26 @@ impl fmt::Debug for InternalRPCHandle {
 
 #[derive(Debug)]
 enum InternalCommand {
-    BasicAck(ChannelId, DeliveryTag, BasicAckOptions, PromiseResolver<()>),
+    BasicAck(
+        ChannelId,
+        DeliveryTag,
+        BasicAckOptions,
+        PromiseResolver<()>,
+        Option<ErrorHolder>,
+    ),
     BasicNack(
         ChannelId,
         DeliveryTag,
         BasicNackOptions,
         PromiseResolver<()>,
+        Option<ErrorHolder>,
     ),
     BasicReject(
         ChannelId,
         DeliveryTag,
         BasicRejectOptions,
         PromiseResolver<()>,
+        Option<ErrorHolder>,
     ),
     CancelConsumer(ChannelId, String, ConsumerStatus),
     CloseChannel(ChannelId, ReplyCode, String),
@@ -214,75 +229,81 @@ impl InternalRPC {
     pub(crate) async fn run(self, channels: Channels) {
         use InternalCommand::*;
 
+        let get_channel = |id| channels.get(id).ok_or(Error::InvalidChannel(id));
+
         while let Ok(Some(command)) = self.rpc.recv_async().await {
             trace!(?command, "Handling internal RPC command");
             let handle = self.handle();
             match command {
-                BasicAck(channel_id, delivery_tag, options, resolver) => channels
-                    .get(channel_id)
-                    .map(|channel| {
-                        self.handle.register_internal_future_with_resolver(
-                            async move { channel.basic_ack(delivery_tag, options).await },
-                            resolver,
-                        )
-                    })
-                    .unwrap_or_default(),
-                BasicNack(channel_id, delivery_tag, options, resolver) => channels
-                    .get(channel_id)
-                    .map(|channel| {
-                        self.handle.register_internal_future_with_resolver(
-                            async move { channel.basic_nack(delivery_tag, options).await },
-                            resolver,
-                        )
-                    })
-                    .unwrap_or_default(),
-                BasicReject(channel_id, delivery_tag, options, resolver) => channels
-                    .get(channel_id)
-                    .map(|channel| {
-                        self.handle.register_internal_future_with_resolver(
-                            async move { channel.basic_reject(delivery_tag, options).await },
-                            resolver,
-                        )
-                    })
-                    .unwrap_or_default(),
-                CancelConsumer(channel_id, consumer_tag, consumer_status) => channels
-                    .get(channel_id)
-                    .map(|channel| {
+                BasicAck(channel_id, delivery_tag, options, resolver, error) => {
+                    let channel = get_channel(channel_id);
+                    self.handle.register_internal_future_with_resolver(
+                        async move {
+                            if let Some(error) = error {
+                                error.check()?;
+                            }
+                            channel?.basic_ack(delivery_tag, options).await
+                        },
+                        resolver,
+                    )
+                }
+                BasicNack(channel_id, delivery_tag, options, resolver, error) => {
+                    let channel = get_channel(channel_id);
+                    self.handle.register_internal_future_with_resolver(
+                        async move {
+                            if let Some(error) = error {
+                                error.check()?;
+                            }
+                            channel?.basic_nack(delivery_tag, options).await
+                        },
+                        resolver,
+                    )
+                }
+                BasicReject(channel_id, delivery_tag, options, resolver, error) => {
+                    let channel = get_channel(channel_id);
+                    self.handle.register_internal_future_with_resolver(
+                        async move {
+                            if let Some(error) = error {
+                                error.check()?;
+                            }
+                            channel?.basic_reject(delivery_tag, options).await
+                        },
+                        resolver,
+                    )
+                }
+                CancelConsumer(channel_id, consumer_tag, consumer_status) => {
+                    let channel = get_channel(channel_id);
+                    handle.register_internal_future(async move {
+                        let channel = channel?;
                         if channel.status().connected() && consumer_status.state().is_active() {
-                            handle.register_internal_future(async move {
-                                channel
-                                    .basic_cancel(&consumer_tag, BasicCancelOptions::default())
-                                    .await
-                            })
+                            channel
+                                .basic_cancel(&consumer_tag, BasicCancelOptions::default())
+                                .await
+                        } else {
+                            Ok(())
                         }
                     })
-                    .unwrap_or_default(),
-                CloseChannel(channel_id, reply_code, reply_text) => channels
-                    .get(channel_id)
-                    .map(|channel| {
-                        handle.register_internal_future(async move {
-                            channel.close(reply_code, &reply_text).await
-                        })
+                }
+                CloseChannel(channel_id, reply_code, reply_text) => {
+                    let channel = get_channel(channel_id);
+                    handle.register_internal_future(async move {
+                        channel?.close(reply_code, &reply_text).await
                     })
-                    .unwrap_or_default(),
-                CloseConnection(reply_code, reply_text, class_id, method_id) => channels
-                    .get(0)
-                    .map(move |channel0| {
-                        handle.register_internal_future(async move {
-                            channel0
-                                .connection_close(reply_code, &reply_text, class_id, method_id)
-                                .await
-                        })
+                }
+                CloseConnection(reply_code, reply_text, class_id, method_id) => {
+                    let channel = get_channel(0);
+                    handle.register_internal_future(async move {
+                        channel?
+                            .connection_close(reply_code, &reply_text, class_id, method_id)
+                            .await
                     })
-                    .unwrap_or_default(),
-                SendConnectionCloseOk(error) => channels
-                    .get(0)
-                    .map(move |channel| {
-                        handle.register_internal_future(async move {
-                            channel.connection_close_ok(error).await
-                        })
+                }
+                SendConnectionCloseOk(error) => {
+                    let channel = get_channel(0);
+                    handle.register_internal_future(async move {
+                        channel?.connection_close_ok(error).await
                     })
-                    .unwrap_or_default(),
+                }
                 RemoveChannel(channel_id, error) => {
                     let channels = channels.clone();
                     handle