+++ src/lib.rs
@@ -146,6 +146,7 @@ mod consumer_status;
 mod consumers;
 mod error;
 mod error_handler;
+mod error_holder;
 mod exchange;
 mod frames;
 mod id_sequence;