+++ src/message.rs
@@ -1,5 +1,6 @@
 use crate::{
     acker::Acker,
+    error_holder::ErrorHolder,
     internal_rpc::InternalRPCHandle,
     protocol::AMQPError,
     types::ShortString,
@@ -60,6 +61,7 @@ impl Delivery {
         routing_key: ShortString,
         redelivered: bool,
         internal_rpc: Option<InternalRPCHandle>,
+        error: Option<ErrorHolder>,
     ) -> Self {
         Self {
             delivery_tag,
@@ -68,7 +70,7 @@ impl Delivery {
             redelivered,
             properties: BasicProperties::default(),
             data: Vec::default(),
-            acker: Acker::new(channel_id, delivery_tag, internal_rpc),
+            acker: Acker::new(channel_id, delivery_tag, internal_rpc, error),
         }
     }
 
@@ -109,6 +111,7 @@ impl BasicGetMessage {
                 routing_key,
                 redelivered,
                 Some(internal_rpc),
+                None,
             ),
             message_count,
         }
@@ -144,7 +147,7 @@ impl BasicReturnMessage {
         reply_text: ShortString,
     ) -> Self {
         Self {
-            delivery: Delivery::new(0, 0, exchange, routing_key, false, None),
+            delivery: Delivery::new(0, 0, exchange, routing_key, false, None, None),
             reply_code,
             reply_text,
         }